requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
idna==3.3
msal==1.17.0
numpy==1.22.3
pandas==1.4.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-power-bi==0.1.2
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9